requirements.txt
opencv-python-headless
requests
numpy
fastai
Pillow
pytubefix
python-dotenv

